Modification required to make bb interface 'live'
kitties.fetch({update:true}) will add/remove/change models to sync the client-side collection to the server. events are triggered on those models.

in kitty_view, view binds to the models' remove and change events with callbacks that remove or render the method.

index_view binds to the collection's 'add' event to render new record.

In this way, the interface will continually update its own collection, triggering events that allow views to keep the overall interface in sync with the server.

NOTE: adding the {wait:true} option to the new view was required because the kitty_view's template currently renders the id of each kitty.